legal_new copy/requirements.txt
# ── Core ML ──
torch>=2.0
transformers>=4.35
sentence-transformers>=2.2
gliner>=0.2
# ── Data pipeline ──
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
feedparser>=6.0
aiohttp>=3.9
pandas>=2.1
numpy>=1.24
jsonlines>=4.0
tqdm>=4.66
# ── Backend ──
fastapi>=0.109
uvicorn[standard]>=0.27
pydantic>=2.5
aiosqlite>=0.19
sqlalchemy[asyncio]>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6
httpx>=0.26
# ── Utilities ──
python-dotenv>=1.0
loguru>=0.7
scikit-learn>=1.3
scipy>=1.11
schedule>=1.2
# ── HuggingFace Datasets ──
datasets>=2.16
# ── Scraper ──
aiohttp>=3.9
